requirements.txt
streamlit==1.56.0
pandas==2.1.4
numpy==1.26.4
plotly==5.18.0
pyteomics==4.7.2
requests==2.31.0
boto3==1.34.0
python-dotenv==1.0.0
loguru==0.7.2
scipy==1.11.4
lxml==5.1.0
sqlalchemy==2.0.25
